float should be int
git-svn-id: https://aften.svn.sourceforge.net/svnroot/aften@168 ef0d8562-5c19-0410-972e-841db63a069c

+++ libaften/exponent.c
@@ -78,7 +78,7 @@ static void
 compute_exponent_strategy(A52Context *ctx)
 {
     int ch, i, j;
-    FLOAT exp_diff;
+    int exp_diff;
     int is_lfe;
     A52Frame *frame;
     A52Block *blocks;
